fix: change nav and share across all pages
