+++ views/components/nav.html
@@ -1,0 +1,9 @@
+{{define "nav.html"}}
+<nav>
+  <a href="/">Home</a>
+  <a href="/about">About</a>
+  <a href="/projects">Projects</a>
+  <a href="/experience">Experience</a>
+  <a href="/blog">Blog</a>
+</nav>
+{{end}}
