+++ views/layouts/blogLayout.html
@@ -8,6 +8,7 @@
   <link href="/static/css/styles.css" rel="stylesheet">
 </head>
 <body>
+  {{template "nav.html" .}}
   <main>
     {{template "blog.html" .blog}}
